rqbalance: Fix uninitialized variable
When looking for idle top/bottom frequencies in the cpufreq
transitioner, the variable to select the cluster number (to
get the idle frequencies related to the cluster) is initialized
only in case we are HMP and want the BIG parameters: this
can lead to various issues, including NULL pointer panic,
and/or random memory address access that may read an abnormally
high or low number and at least make the algo to not work as
expected.

Initialize the uninitialized variable as zero, so that the
default cluster is LITTLE, as per initial design.

Signed-off-by: Ad!thya R <adithya.r02@outlook.com>